+++ src/jaeger/alerts.py
@@ -370,6 +370,10 @@ class AlertsBot(BaseBot):
     async def _check_chiller(self):
         """Checks the chiller status."""
 
+        # Fluid temperature alert disabled on 2024-06-15 after discussion
+        # in the opscall. We agreed that this alert is not critical and we were not
+        # sure how we'd act in case it's triggered.
+
         if not isinstance(self.ieb, IEB):
             self.notify("IEB not connected. Cannot run chiller checks.")
             return
@@ -378,7 +382,7 @@ class AlertsBot(BaseBot):
         assert chiller is not None
 
         try:
-            setpoint = (await chiller.read_device("TEMPERATURE_USER_SETPOINT"))[0]
+            # setpoint = (await chiller.read_device("TEMPERATURE_USER_SETPOINT"))[0]
             fluid_temp = (await chiller.read_device("DISPLAY_VALUE"))[0]
         except Exception as err:
             self.notify(f"Failed reading chiller values: {err}", level=logging.ERROR)
@@ -392,16 +396,16 @@ class AlertsBot(BaseBot):
         else:
             self.set_keyword("alert_chiller_dew_point", False)
 
-        chiller_config = config["alerts"]["chiller"]
+        # chiller_config = config["alerts"]["chiller"]
 
-        supply_temp = (await self.ieb.read_device(chiller_config["sensor_supply"]))[0]
+        # supply_temp = (await self.ieb.read_device(chiller_config["sensor_supply"]))[0]
 
-        if abs(setpoint - supply_temp) > chiller_config["threshold"]:
-            self.set_keyword("alert_fluid_temperature", True)
-            self.notify("Chiller set point is different from supply temperature.")
+        # if abs(setpoint - supply_temp) > chiller_config["threshold"]:
+        #     self.set_keyword("alert_fluid_temperature", True)
+        #     self.notify("Chiller set point is different from supply temperature.")
 
-        else:
-            self.set_keyword("alert_fluid_temperature", False)
+        # else:
+        #     self.set_keyword("alert_fluid_temperature", False)
 
         # Check if there are chiller alerts.
         chiller_alerts: list[str] = []